api/requirements.txt
bottle==0.12.13
requests==2.20.0
python-dateutil==2.7.5
gunicorn==19.9.0
